refactor: todo render 함수들 정리, 모듈 정리
