+++ src/Team/removeTeam.js
@@ -1,11 +1,10 @@
 import { clearAllList, getTeamId } from './team.js';
 import { loadTeamList } from './loadTeam.js';
 import { api } from '../api.js';
+import { showMessage } from '../validator.js';
 
 export const removeTeam = async (target) => {
-  if (!confirm('정말 이 팀을 삭제하시겠습니까?')) {
-    return;
-  }
+  showMessage('정말 이 팀을 삭제하시겠습니까?');
 
   const teamId = getTeamId(target);
