--- src/Team/teamTitle.js
@@ -1,1 +1,0 @@
-// 생각해보니 타이틀에 팀 이름도 바뀌어야함
