+++ src/Todo/feature/todoInput.js
@@ -1,24 +1,20 @@
-import { api } from '../api.js';
-import { template } from '../template.js';
-import { teamId, getMemberId } from './todo.js';
+import { api } from '../../etc/api.js';
+import { loadTodo } from '../loadTodo.js';
+import { teamId, getMemberId } from '../todo.js';
 
 export const todoInput = ({ target, key }) => {
   if (!target.classList.contains('new-todo')) return;
   if (!target.value || key !== 'Enter') return;
 
   const addTodo = async () => {
-    const $todoList = target.closest('div').querySelector('.todo-list');
     const memberId = getMemberId(target);
-
     const todoContents = {
       contents: target.value,
     };
 
-    $todoList.insertAdjacentHTML('beforeend', template.todoItem(target.value, ''));
-
     await api.addTodo(teamId, memberId, todoContents);
+    loadTodo();
   };
 
   addTodo();
-  target.value = '';
 };
