+++ src/etc/dom.js
