+++ src/etc/validator.js
@@ -1,0 +1,5 @@
+export const showMessage = (string) => {
+  if (!confirm(string)) {
+    return;
+  }
+};
