--- src/main.js
@@ -1,1 +1,0 @@
-// 메인에서 team.js랑 todo.js 불러와서 실행시키려니까 화면이 다를땐 에러가 뜸
